+++ lib/src/util/logger.dart
@@ -9,7 +9,7 @@ var _logger = Logger(
   level: kReleaseMode ? Level.info : Level.debug,
   printer: PrettyPrinter(
       stackTraceBeginIndex: 1,
-      methodCount: 2,
+      methodCount: 5,
       errorMethodCount: 8,
       lineLength: 120,
       colors: true,