+++ lib/src/util/num_util.dart
@@ -1,0 +1,9 @@
+extension NumUtil on num {
+  String format() {
+    if (isInfinite) {
+      return toStringAsFixed(2);
+    } else {
+      return truncateToDouble() == this ? toStringAsFixed(0) : toString();
+    }
+  }
+}