+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Add json data model to storage key, value and comment;
+- New api;
+- fix [Editor not updating · Issue #1 · youngchan1988/json_editor (github.com)](https://github.com/youngchan1988/json_editor/issues/1)
+
 ## 0.0.1
 
 Initial project. Support the follow features: